requirements-prod.txt
# Runtime
PyNaCl==1.5.0
